feat(FE): 리스트뷰 기능별 세분화 및 임시 프로필 이미지 추가 BUS-191
